feat: moved category ratings charts to dashboard page
