+++ src/pages/Home/index.tsx
@@ -1,20 +1,16 @@
 import { Container, InputsContainer } from './styles.ts';
 import BooksTable from '../../components/Table';
 import { useEffect, useState } from 'react';
-import { getBooksByCategory, getByAuthorAndTitle, getVolumes } from '../../services/GoogleBooksService.ts';
+import { getByAuthorAndTitle, getVolumes } from '../../services/GoogleBooksService.ts';
 import { APIBook } from '../../types/apiTypes.ts';
-import { Book, CategoryRatings } from '../../types/commonTypes.ts';
+import { Book } from '../../types/commonTypes.ts';
 import SearchFilter from '../../components/SearchFilter';
-import RatingsTimeline from '../../components/RatingsTimeline';
 import { deferFunction } from '../../utils/deferFunction.ts';
-import CategoriesChart from '../../components/CategoriesChart';
 
 export default function Home() {
   const [books, setBooks] = useState<Book[]>([]);
   const [author, setAuthor] = useState<string>('');
   const [title, setTitle] = useState<string>('');
-  const [categoryRatings, setCategoryRatings] = useState<CategoryRatings[]>([]);
-  const exampleCategories = ['fiction', 'romance', 'medical', 'engineering', 'cooking'];
 
   function onTitleChange(title: string): void {
     setTitle(title);
@@ -41,27 +37,6 @@ export default function Home() {
       setBooks(data);
     }
   }
-
-  useEffect(() => {
-    async function getRatingsByCategories(categories: string[]) {
-      for (const category of categories) {
-        const response = await getBooksByCategory({ category: category, orderBy: 'newest' });
-        if (response !== undefined) {
-          const rates = response.items
-            .filter((item) => !!item.volumeInfo.averageRating)
-            .map((item) => {
-              return {
-                category: category,
-                rating: item.volumeInfo.averageRating,
-              };
-            });
-          setCategoryRatings((prevState) => [...prevState].concat(rates));
-        }
-      }
-    }
-
-    getRatingsByCategories(exampleCategories);
-  }, []);
 
   useEffect(() => {
     async function getBooks() {
@@ -100,8 +75,6 @@ export default function Home() {
         />
       </InputsContainer>
       <BooksTable books={books} />
-      <CategoriesChart categories={exampleCategories} ratingsData={categoryRatings} />
-      <RatingsTimeline categories={exampleCategories} ratings={categoryRatings} />
     </Container>
   );
 }
